Add data container filter and filter chain
Add an interface for data container filters, to be able to filter data
containers based on their contents.

The filter is implemented by a filter chain, so as to be able to both
chain multiple filters, as well as have a default filter that always
returns true if left empty.